// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AdminRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  MODERATOR: 'MODERATOR',
  SUPPORT: 'SUPPORT'
} as const

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  BANNED: 'BANNED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const ReportStatus = {
  PENDING: 'PENDING',
  INVESTIGATING: 'INVESTIGATING',
  RESOLVED: 'RESOLVED',
  DISMISSED: 'DISMISSED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const ReportedType = {
  USER: 'USER',
  HOST: 'HOST',
  SIMULATOR: 'SIMULATOR',
  REVIEW: 'REVIEW'
} as const

export type ReportedType = (typeof ReportedType)[keyof typeof ReportedType]


export const UserType = {
  CUSTOMER: 'CUSTOMER',
  HOST: 'HOST',
  ADMIN: 'ADMIN'
} as const

export type UserType = (typeof UserType)[keyof typeof UserType]